Refactoring

+++ 2.JavaCore/src/com/javarush/task/task15/task1506/Solution.java
@@ -10,7 +10,7 @@ public class Solution {
         print((Number) 1);
         print(1);
         print((Integer) 1);
-        print((int) 1);
+        print(1);
     }
 
     public static void print(Integer i) {
